update the way generating 2d random numbers; now all random# rely on the function Sample()

+++ sbncode/SBNEventWeight/Calculators/BNBFlux/PrimaryHadronFeynmanScalingWeightCalc.cxx
@@ -6,7 +6,7 @@
 namespace sbn {
 	namespace evwgh {
 
-		std::pair<bool, double> FluxWeightCalc::PHFSWeightCalc(simb::MCFlux flux, std::vector<double> rand){
+		std::pair<bool, double> FluxWeightCalc::PHFSWeightCalc(simb::MCFlux flux, std::vector<float> rand){
 
 			// 
 			//  Largely built off the MiniBooNE code
